refactor(api): convert all model IDs from number to uuid
