+++ packages/api/src/auth/jwt/jwt-user-id.decorator.ts
@@ -6,7 +6,7 @@ import type { JwtAuthGuardReq } from './jwt.types';
  * Returns id of the authorized user.
  * @type param decorator
  */
-export const UserId = createParamDecorator((_data: unknown, ctx: ExecutionContext) => {
+export const JwtUserId = createParamDecorator((_data: unknown, ctx: ExecutionContext) => {
   const request = ctx.switchToHttp().getRequest<JwtAuthGuardReq>();
 
   return request.user.id;
