+++ packages/api/src/learning-materials/ports/users.port.ts
@@ -1,5 +1,7 @@
+import type { UserId } from '../../users/users.types';
+
 export const USERS_PORT = Symbol('USERS_PORT');
 
 export interface UsersPort {
-  getUserFullNameById(userId: number): Promise<string>;
+  getUserFullNameById(userId: UserId): Promise<string>;
 }
